requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
boto3==1.38.46
botocore==1.38.46
celery==5.5.3
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.13.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
kombu==5.5.4
MarkupSafe==3.0.2
openai==1.93.0
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-http-client==3.3.7
redis==6.2.0
requests==2.32.4
s3transfer==0.13.0
sendgrid==6.12.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==12.2.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
